requirement.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
oauthlib==3.2.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
tornado==6.1
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
